% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_pf_across_polygons.R
\name{aggregate_pf_across_polygons}
\alias{aggregate_pf_across_polygons}
\title{Map point observations to polygons and aggregate them across user-specified grouping variables.}
\usage{
aggregate_pf_across_polygons(
  data,
  polygons,
  crs,
  group_by_variables = c("polygon_id", "longitude", "latitude", "locus")
)
}
\arguments{
\item{data}{Input data object.}

\item{polygons}{An `sf` object containing aggregation polygons.}

\item{crs}{Object supplying the coordinate reference system for input points.}

\item{group_by_variables}{Character vector of columns used to group observations during aggregation.}
}
\value{
The result produced by the function.@
}
\description{
Map point observations to polygons and aggregate them across user-specified grouping variables.
}
