% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_pf_data_in_polygons.R
\name{aggregate_pf_data_in_polygons}
\alias{aggregate_pf_data_in_polygons}
\title{aggregate_pf_data_in_polygons}
\usage{
aggregate_pf_data_in_polygons(data, polygons, polygon_id_column)
}
\description{
aggregate_pf_data_in_polygons
}
