% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.excluded.R
\name{check.excluded}
\alias{check.excluded}
\title{Identify spatial observations that lie outside continental landmasses}
\usage{
check.excluded(data_sf, continents_sf)
}
\arguments{
\item{data_sf}{An `sf` object containing the observations.}

\item{continents_sf}{An `sf` object containing continent polygons.}
}
\value{
A list containing included and excluded observations together with the original spatial inputs.@
}
\description{
Identify spatial observations that lie outside continental landmasses
}
