% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.HbS.prediction.extent.R
\name{compute.HbS.prediction.extent}
\alias{compute.HbS.prediction.extent}
\title{Construct the spatial extent used for HbS predictions from a reference raster and selected country boundaries.}
\usage{
compute.HbS.prediction.extent(
  world_sf,
  map_filename =
    "geodata/2013_Sickle_Haemoglobin_HbS_Allele_Freq_Global_5k_Decompressed.tif",
  notpiel = 0.005
)
}
\arguments{
\item{world_sf}{World country polygons as an `sf` object.}

\item{map_filename}{Path to the reference HbS raster used to define the prediction extent.}

\item{notpiel}{Threshold used to define areas retained from HbS by Piel et al. (benchmark).}
}
\value{
The result produced by the function.@
}
\description{
Construct the spatial extent used for HbS predictions from a reference raster and selected country boundaries.
}
