% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_scientific_to_numeric.R
\name{convert_scientific_to_numeric}
\alias{convert_scientific_to_numeric}
\title{convert_scientific_to_numeric}
\usage{
convert_scientific_to_numeric(x)
}
\description{
convert_scientific_to_numeric
}
