% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.colours.R
\name{country.colours}
\alias{country.colours}
\title{Return the named colour palette used for countries in manuscript figures.}
\usage{
country.colours()
}
\value{
A named character vector of colours.@
}
\description{
Return the named colour palette used for countries in manuscript figures.
}
