% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key_hex_custom.R
\name{draw_key_hex_custom}
\alias{draw_key_hex_custom}
\title{Draw a custom hexagonal ggplot2 legend key.}
\usage{
draw_key_hex_custom(data, params, size)
}
\arguments{
\item{data}{Input data object.}

\item{params}{Additional parameters supplied by ggplot2 when drawing the legend key.}

\item{size}{Legend key size.}
}
\value{
The result produced by the function.@
}
\description{
Draw a custom hexagonal ggplot2 legend key.
}
