% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HbSplottheme.R
\docType{data}
\name{HbSplottheme}
\alias{HbSplottheme}
\title{HbSplottheme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 7.
}
\usage{
HbSplottheme
}
\description{
HbSplottheme
}
\keyword{datasets}
