% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.logit.R
\name{inverse.logit}
\alias{inverse.logit}
\title{Apply the inverse-logit transformation.}
\usage{
inverse.logit(x)
}
\arguments{
\item{x}{Input value or vector.}
}
\value{
A numeric vector on the probability scale.@
}
\description{
Apply the inverse-logit transformation.
}
