% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.genotypes.from.vcf.R
\name{load.genotypes.from.vcf}
\alias{load.genotypes.from.vcf}
\title{Load genotypes and read counts for specified variants from a VCF file}
\usage{
load.genotypes.from.vcf(filename, variants)
}
\arguments{
\item{filename}{Path to an input file.}

\item{variants}{A dataframe of variants with chromosome and position columns}
}
\value{
a data frame of samples, genotypes, and ref and alt read counts
}
\description{
Load genotypes and read counts for specified variants from a VCF file,
Uses bcftools query under the hood; this must be installed.
}
