% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_HbS_mean.R
\name{load_HbS_mean}
\alias{load_HbS_mean}
\title{Load an HbS mean prediction raster and prepare it for downstream use.}
\usage{
load_HbS_mean(filename)
}
\arguments{
\item{filename}{Path to an input file.}
}
\value{
The result produced by the function.@
}
\description{
Load an HbS mean prediction raster and prepare it for downstream use.
}
