% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Fit a logistic model to the supplied data using the requested formula.}
\usage{
logistic(data, formula = Y ~ year)
}
\arguments{
\item{data}{Input data object.}

\item{formula}{Model formula.}
}
\value{
The result produced by the function.@
}
\description{
Fit a logistic model to the supplied data using the requested formula.
}
