% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdir_recursive.R
\name{mkdir_recursive}
\alias{mkdir_recursive}
\title{Create a directory and any missing parent directories.}
\usage{
mkdir_recursive(path)
}
\arguments{
\item{path}{Directory path to create.}
}
\value{
A logical value returned invisibly by `dir.create()`.@
}
\description{
Create a directory and any missing parent directories.
}
