% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_sample_genotypes_to_db.R
\name{output_sample_genotype_to_db}
\alias{output_sample_genotype_to_db}
\title{Output sample genotypes and read counts to a sqlite database, for downstream analysis}
\usage{
output_sample_genotype_to_db(by_sample, source, filename)
}
\arguments{
\item{by_sample}{data framt of sample IDs, genotypes, and read counts for a set of variants}

\item{source}{a name to add in a 'source' column in the output}

\item{filename}{filename of sqlite db to output to}
}
\description{
Output sample genotypes and read counts to a sqlite database, for downstream analysis
}
