% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_inla_binomial_model.R
\name{predict_inla_binomial_model}
\alias{predict_inla_binomial_model}
\title{Generate posterior predictions from an INLA binomial spatial model and summarise their distributions.}
\usage{
predict_inla_binomial_model(
  posterior.samples,
  mesh,
  prediction_locations,
  covariates = NULL
)
}
\arguments{
\item{posterior.samples}{List of posterior samples returned by INLA.}

\item{mesh}{INLA mesh object.}

\item{prediction_locations}{Matrix or data frame of coordinates at which to predict.}

\item{covariates}{Optional data frame or matrix of covariate values.}
}
\value{
A list containing posterior predictions and their mean, standard deviation, quartiles, and interquartile range.@
}
\description{
Generate posterior predictions from an INLA binomial spatial model and summarise their distributions.
}
