% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_polygons.R
\name{aggregate_to_polygons}
\alias{aggregate_to_polygons}
\title{aggregate_to_polygons}
\usage{
aggregate_to_polygons(data, countries, polygons, polygon_id = "NAME_2")
}
\description{
aggregate_to_polygons
}
