% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.continent.covariates.R
\name{build.continent.covariates}
\alias{build.continent.covariates}
\title{build.continent.covariates}
\usage{
build.continent.covariates(sf, world_sf)
}
\description{
build.continent.covariates
}
