% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.HbS.prediction.extent.R
\name{compute.HbS.prediction.extent}
\alias{compute.HbS.prediction.extent}
\title{compute.HbS.prediction.extent}
\usage{
compute.HbS.prediction.extent(
  world_sf,
  map_filename =
    "geodata/2013_Sickle_Haemoglobin_HbS_Allele_Freq_Global_5k_Decompressed.tif",
  notpiel = 0.005
)
}
\description{
compute.HbS.prediction.extent
}
