% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig1b.plot.R
\name{fig1b.plot}
\alias{fig1b.plot}
\title{fig1b.plot}
\usage{
fig1b.plot(
  pfpt,
  border,
  scicopalette,
  savepath,
  allele = NULL,
  myheight = myheight,
  mywidth = mywidth,
  myproj = NA
)
}
\description{
fig1b.plot
}
