% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_inla_binomial_model.R
\name{fit_inla_binomial_model}
\alias{fit_inla_binomial_model}
\title{fit_inla_binomial_model}
\usage{
fit_inla_binomial_model(xyt, extpoly, prior, covariate = NULL, verbose = FALSE)
}
\description{
fit_inla_binomial_model
}
