% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diagnostic_plot.R
\name{generate_diagnostic_plot}
\alias{generate_diagnostic_plot}
\title{generate_diagnostic_plot}
\usage{
generate_diagnostic_plot(
  xyt,
  modelfit,
  predictions,
  HbSPiel,
  features = list(spatialdomain = africa_sf, rivers = rivaf_sf, lakes = lakaf_sf),
  color.scheme,
  titles = list(t1 = "HbS | Predicted mean prevalence", t2 =
    "HbS | Predicted standard deviation", t3 = "HbS | Predicted Q25", t4 =
    "HbS | Predicted Q75", t5 = "HbS | Predicted IQR"),
  prednames = c("mean", "sd", "iqr"),
  popmask,
  saveraster,
  saverastername = "HbS"
)
}
\description{
generate_diagnostic_plot
}
