% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_raster_maps.R
\name{generate_raster_maps}
\alias{generate_raster_maps}
\title{generate_raster_maps}
\usage{
generate_raster_maps(
  predictions,
  saveraster = FALSE,
  saverastername = saverastername,
  savepath = "output/HbSraster/"
)
}
\description{
generate_raster_maps
}
