% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_locations.R
\name{get_prediction_locations}
\alias{get_prediction_locations}
\title{get_prediction_locations}
\usage{
get_prediction_locations(alt, study_area, masked_features = list())
}
\description{
get_prediction_locations
}
