% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.continent.shapes.terra.R
\name{load.continent.shapes.terra}
\alias{load.continent.shapes.terra}
\title{load.continent.shapes.terra}
\usage{
load.continent.shapes.terra(filename, continent = NA)
}
\description{
load.continent.shapes.terra
}
