% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.piel_et_al_data.R
\name{load.piel_et_al_data}
\alias{load.piel_et_al_data}
\title{load.piel_et_al_data}
\usage{
load.piel_et_al_data(
  filename,
  exclude_non_mh = FALSE,
  exclude_wide_area = FALSE
)
}
\description{
load.piel_et_al_data
}
