% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeinlastack.binomial.R
\name{makeinlastack.binomial}
\alias{makeinlastack.binomial}
\title{makeinlastack.binomial}
\usage{
makeinlastack.binomial(Y, n, A, spde, covariate = NULL)
}
\description{
makeinlastack.binomial
}
