% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_inla_binomial_model.R
\name{predict_inla_binomial_model}
\alias{predict_inla_binomial_model}
\title{predict_inla_binomial_model}
\usage{
predict_inla_binomial_model(
  posterior.samples,
  mesh,
  prediction_locations,
  covariates = NULL
)
}
\description{
predict_inla_binomial_model
}
