% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_values.R
\name{predict_values}
\alias{predict_values}
\title{predict_values}
\usage{
predict_values(
  nn,
  posterior.samples,
  A.pred,
  covariates = NULL,
  link.function = stats::plogis
)
}
\description{
predict_values
}
