% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runinla.binomial.R
\name{runinla.binomial}
\alias{runinla.binomial}
\title{runinla.binomial}
\usage{
runinla.binomial(
  myformula,
  stk,
  spde,
  n,
  covariate.prec = 0.001,
  intercept.prec = 0
)
}
\description{
runinla.binomial
}
