% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.panelgrid.R
\docType{data}
\name{theme.panelgrid}
\alias{theme.panelgrid}
\title{theme.panelgrid}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 1.
}
\usage{
theme.panelgrid
}
\description{
theme.panelgrid
}
\keyword{datasets}
