% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridisoption.R
\docType{data}
\name{viridisoption}
\alias{viridisoption}
\title{viridisoption}
\format{
An object of class \code{list} of length 2.
}
\usage{
viridisoption
}
\description{
viridisoption
}
\keyword{datasets}
